Add basic feature test
Adds a simple test to ensure that the basic functionality of the application is working correctly, and we're able to serve content from a basic route.

This test will probably be removed later, but ensures that we haven't broken any core functionality until we get more meaningful code and tests in place.